refactor: extract mobile menu and gallery components
