+++ src/components/header/HeaderMiddle.tsx
@@ -1,5 +1,5 @@
 import React from "react"
-import { default as skyseeLogo } from "/src/assets/skysee-7.png"
+import { Logo } from "./Logo"
 import { WindowSize } from "../../containers/WindowSize"
 import { Link } from "react-router-dom"
 
@@ -12,7 +12,7 @@ export const HeaderMiddle: React.FC = () => {
           to="/"
           className="absolute left-1/2 top-1/2 -translate-x-1/2 -translate-y-1/2 py-1 z-10"
         >
-          <img src={skyseeLogo} className="h-12 w-auto" />
+          <Logo className="h-12 w-auto" />
         </Link>
       )}
     </div>
